# Conflicts:
#	app/build.gradle

+++ app/src/main/java/com/wallet/crypto/trustapp/repository/TokenRepository.java
@@ -29,7 +29,6 @@ import org.web3j.protocol.http.HttpService;
 import org.web3j.utils.Numeric;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -302,8 +301,8 @@ public class TokenRepository implements TokenRepositoryType {
         return response.getValue();
     }
 
-    public static byte[] createTokenTransferData(String to, BigInteger tokenAmount) {
-        List<Type> params = Arrays.asList(new Address(to), new Uint256(tokenAmount));
+    public static byte[] createTokenTransferData(String to, BigDecimal tokenAmount) {
+        List<Type> params = Arrays.asList(new Address(to), new Uint256(tokenAmount.toBigInteger()));
         List<TypeReference<?>> returnTypes = Collections.singletonList(new TypeReference<Bool>() {});
         Function function = new Function("transfer", params, returnTypes);
         String encodedFunction = FunctionEncoder.encode(function);
